fixed a couple stupid mistakes
git-svn-id: https://bugengine.googlecode.com/svn/trunk@379 f3700e90-1ba7-11de-b56b-c9554f25afe1

+++ src/engine/core/api/core/string/istring.hh
@@ -136,7 +136,7 @@ be_api(CORE) ifilename operator+(const ipath& path, const ifilename& file);
 
 namespace minitl
 {
-inline size_t hash_value(const BugEngine::istring& key) { return key.hash(); }
+inline size_t hash_value(const BugEngine::istring& key) { return (size_t)key.hash(); }
 
 template< u16 size >
 const format<size>& operator|(const format<size>& f, const BugEngine::istring& value)
